add http stub/mock server
the server can match on routes, http method, sent payload from
client. it is also possible to match by custom assertion on
request, response and sent data. when a matching stub/mock
is found, there can be optional testing, too.

you can pass a mock definition, a list of mock definition or
a function (returning you own mock middleware) to the server.

for many use cases a simple match by route and a simple reply is
ok, so we have a convenience shortshand for this.

see example-server-simple.js and example-server-advanced.js
for each usage.